chore: move GetTitleMeta to utils
