+++ gamescollection/collection.go
@@ -11,7 +11,6 @@ import (
 	"fmt"
 	"io"
 	"log"
-	"math/big"
 	"os"
 	"path/filepath"
 	"strings"
@@ -223,7 +222,7 @@ func (c *collect) AddNewGames(newGames []repository.FileDesc) {
 	var gameList = make([]repository.GameFileType, 0)
 
 	for _, file := range newGames {
-		baseID, update, dlc := GetTitleMeta(file.GameID)
+		baseID, update, dlc := utils.GetTitleMeta(file.GameID)
 		baseTitle := c.Library()[baseID]
 		title := c.Library()[file.GameID]
 		extension := filepath.Ext(file.Path)
@@ -281,42 +280,4 @@ func (c *collect) GetKey(gameID string) (string, error) {
 		return "", errors.New("TitleDBKey for game " + gameID + " is not found")
 	}
 	return string(key), nil
-}
-
-// GetTtitleMeta returns the BaseID of the content, as well as Update / DLC flags
-func GetTitleMeta(titleID string) (string, bool, bool) {
-	var lastDigit = titleID[len(titleID)-1:]
-	var baseID = strings.Join([]string{titleID[:len(titleID)-3], "000"}, "")
-	var update = false
-	var dlc = false
-
-	if titleID != baseID {
-		update = true
-	}
-
-	if lastDigit != "0" {
-		dlc = true
-		update = false
-
-		// Parse the hexadecimal string into a big integer
-		intValue, success := new(big.Int).SetString(baseID, 16)
-		if !success {
-			return "", false, false
-		}
-
-		// Parse the subtraction value (in hexadecimal)
-		subtractionValue := new(big.Int)
-		subtractionValue, success = subtractionValue.SetString("1000", 16)
-		if !success {
-			return "", false, false
-		}
-
-		// Subtract the values
-		intValue.Sub(intValue, subtractionValue)
-
-		// Convert the resulting integer back to a hexadecimal string, left padded with 0 to 16 chars
-		baseID = fmt.Sprintf("0000000000000000%X", intValue)
-		baseID = baseID[len(baseID)-16:]
-	}
-	return baseID, update, dlc
 }
